fix: on_voice_state_update()

+++ bot.py
@@ -96,13 +96,12 @@ def main():
     
     @bot.event
     async def on_voice_state_update(member, before, after):
-        if after.channel is None: 
-            if len(before.channel.members) == 1: 
-                voice_client = discord.utils.get(bot.voice_clients, guild=before.channel.guild)
-                if voice_client:
-                    await guild_controller[voice_client.guild].message.delete()       
-                    guild_controller[voice_client.guild].resetting()
-                    await voice_client.disconnect()
+        if after.channel is None and len(before.channel.members) == 1: 
+            voice_client = discord.utils.get(bot.voice_clients, guild=before.channel.guild)
+            if voice_client and before.channel.id == voice_client.channel.id:
+                await guild_controller[voice_client.guild].message.delete()       
+                guild_controller[voice_client.guild].resetting()
+                await voice_client.disconnect()
 
     @bot.tree.command(name="pin")
     async def pin(interaction: discord.Interaction):
